pyproject.toml
[tool.poetry]
name = "speak2me-fitness-monorepo"
version = "0.1.0"
description = ""
authors = ["stazman <cadistasio@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
pre-commit = "^3.8.0"
flake8 = "^7.1.1"
fastapi = "^0.115.0"
uvicorn = ">=0.31.0,<0.36.0"
motor = "^3.6.0"
beanie = "^1.27.0"
openai = "^2.29.0"
python-multipart = "^0.0.22"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
coverage = "^7.6.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
